requirements.txt
yt-dlp
openai-whisper
torch
torchaudio
ffmpeg-python
tkinter-tooltip
flask
openai
